+++ src/GUI/AndroidBall.hpp
@@ -4,7 +4,7 @@
 
 using namespace geode::prelude;
 
-class AndroidBall : public CCNode
+class AndroidBall : public CCNodeRGBA
 {
     protected:
         CCSprite* background = nullptr;
@@ -13,6 +13,11 @@ class AndroidBall : public CCNode
         bool isDragging = false;
         bool isMoving = false;
         CCPoint moveOffset = ccp(0, 0);
+        CCPoint position = CCPointZero;
+
+        float scale = 1.0f;
+        float normalOpacity = 0.4f;
+        bool smoothMove = true;
         
     public:
         CREATE_FUNC(AndroidBall);
@@ -24,11 +29,15 @@ class AndroidBall : public CCNode
         void reloadTextures();
         void onOpenMenu();
 
+        void setButtonScale(float scale);
+        void setSmoothMove(bool smooth);
+
         bool ccTouchBegan(CCTouch* touch);
         bool ccTouchMoved(CCTouch* touch);
         bool ccTouchEnded(CCTouch* touch);
 
-        virtual void update(float dt);
         virtual bool init();
+        virtual void setOpacity(GLubyte opacity);
+        virtual void update(float dt);
         virtual void visit();
 };